+++ examples/Shmup.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE FlexibleContexts, DataKinds, ScopedTypeVariables, TypeApplications, TypeFamilies, MultiParamTypeClasses, TemplateHaskell #-}
 
 import Apecs
+import Data.Proxy
 import Control.Monad
 import Linear
 import Graphics.Gloss
@@ -40,7 +41,7 @@ type Movable = (Position, Velocity)
 
 playerSpeed = 170
 playerPos = V2 0 (-120)
-scorePos = V2 xmin (-170)
+scorePos  = V2 xmin (-170)
 bulletSpeed = 500
 enemySpeed  = 80
 xmin = -100
@@ -104,7 +105,6 @@ draw = do
   pt <- toPic $ \(Particle col _, Position p, Velocity (V2 vx vy)) -> color col . translate' p $ Line [(0,0),(realToFrac vx/10, realToFrac vy/10)]
   return $ mconcat [p,t,b,s,pt]
   where
-    toPic :: Has World c => (c -> Picture) -> System' Picture
     toPic f = mconcat . fmap f <$> getAll
     translate' (V2 x y) = translate (realToFrac x) (realToFrac y)
     triangle = Line [(0,0),(-0.5,-1),(0.5,-1),(0,0)]