+++ src/Apecs/Core.hs
@@ -17,45 +17,56 @@ import qualified Data.Vector.Unboxed   as U
 import qualified Apecs.THTuples        as T
 
 -- | An Entity is just an integer, used to index into a component store.
+--   In general, use @newEntity@, @cmap@, and component tags instead of manipulating these directly.
+--   
+--   For performance reasons, negative values like (-1) are reserved for stores to represent special values, so avoid using these.
 newtype Entity = Entity {unEntity :: Int} deriving (Num, Eq, Ord, Show)
 
 -- | A System is a newtype around `ReaderT w IO a`, where `w` is the game world variable.
+--   Systems mainly serve to
+--   
+--   * Lift side effects into the IO Monad.
+--
+--   * Allow type-based lookup of a component's store through @getStore@.
 newtype System w a = System {unSystem :: ReaderT w IO a} deriving (Functor, Monad, Applicative, MonadIO)
 
--- | A component is defined by the type of its storage
---   The storage in turn supplies runtime types for the component.
---   For the component to be valid, its Storage must be an instance of Store.
+-- | A component is defined by specifying how it is stored.
+--   The constraint ensures that stores and components are mapped one-to-one.
 class (Elem (Storage c) ~ c) => Component c where
   type Storage c
 
--- | A world `Has` a component if it can produce its Storage
+-- | @Has w c@ means that world @w@ can produce a @Storage c@.
 class Component c => Has w c where
   getStore :: System w (Storage c)
 
--- | The type of components stored by a Store
+-- | The type of components stored by a store, e.g. @Elem (Map c) = c@.
 type family Elem s
 
--- | Holds components indexed by entities
+-- | Indicates that the store @s@ can be initialized.
+--   Generally, \"base\" stores like @Map c@ can be initialized, but composite stores like @MaybeStore s@ cannot.
 class ExplInit s where
-  -- | Initialize the store with its initialization arguments.
+  -- | Initialize a new empty store.
   explInit :: IO s
 
--- | Stores that support @get@ and @exists@ in the IO monad
---   If @existsIO@
+-- | Stores that we can read using @explGet@ and @explExists@.
+--   For some entity @e@, @eplGet s e@ is only guaranteed to be safe if @explExists s e@ returns @True@.
 class ExplGet s where
-  -- | Reads a component from the store. What happens if the component does not exist is left undefined.
+  -- | Reads a component from the store. What happens if the component does not exist is left undefined, and might not necessarily crash.
   explGet :: s -> Int -> IO (Elem s)
-  -- | Returns whether there is a component for the given index
+  -- | Returns whether there is a component for the given index.
   explExists :: s -> Int -> IO Bool
 
+-- | Stores that can be written.
 class ExplSet s where
-  -- | Writes a component
+  -- | Writes a component to the store.
   explSet :: s -> Int -> Elem s -> IO ()
 
+-- | Stores that components can be removed from.
 class ExplDestroy s where
   -- | Destroys the component for a given index.
   explDestroy :: s -> Int -> IO ()
 
+-- | Stores that we can request a list of member entities for.
 class ExplMembers s where
   -- | Returns an unboxed vector of member indices
   explMembers :: s -> IO (U.Vector Int)
@@ -65,6 +76,7 @@ type Set     w c = (Has w c, ExplSet     (Storage c))
 type Members w c = (Has w c, ExplMembers (Storage c))
 type Destroy w c = (Has w c, ExplDestroy (Storage c))
 
+-- | Identity component/store. @Identity c@ is equivalent to @c@, so using it is mostly useless.
 instance Component c => Component (Identity c) where
   type Storage (Identity c) = Identity (Storage c)
 
@@ -84,14 +96,14 @@ instance ExplMembers s => ExplMembers (Identity s) where
 instance ExplDestroy s => ExplDestroy (Identity s) where
   explDestroy (Identity s) = explDestroy s
 
--- Tuple Instances TODO
 T.makeInstances [2..8]
 
-
 -- | Psuedocomponent indicating the absence of @a@.
+--   Mainly used as e.g. @cmap $ \(a, Not b) -> c@ to iterate over entities with an @a@ but no @b@.
+--   Can also be used to delete components, like @cmap $ \a -> (Not :: Not a)@ to delete every @a@ component.
 data Not a = Not
 
--- | Pseudostore used to produce values of type @Not a@
+-- | Pseudostore used to produce values of type @Not a@, inverts @explExists@, and destroys instead of @explSet@.
 newtype NotStore s = NotStore s
 
 instance Component c => Component (Not c) where
@@ -109,7 +121,9 @@ instance ExplGet s => ExplGet (NotStore s) where
 instance ExplDestroy s => ExplSet (NotStore s) where
   explSet (NotStore sa) ety _ = explDestroy sa ety
 
--- | Pseudostore used to produce values of type @Maybe a@
+-- | Pseudostore used to produce values of type @Maybe a@.
+--   Will always return @True@ for @explExists@.
+--   Writing can both set and delete a component using @Just@ and @Nothing@ respectively.
 newtype MaybeStore s = MaybeStore s
 instance Component c => Component (Maybe c) where
   type Storage (Maybe c) = MaybeStore (Storage c)
@@ -130,7 +144,13 @@ instance (ExplDestroy s, ExplSet s) => ExplSet (MaybeStore s) where
   explSet (MaybeStore sa) ety Nothing  = explDestroy sa ety
   explSet (MaybeStore sa) ety (Just x) = explSet sa ety x
 
+-- | Pseudocomponent that functions normally for @explExists@ and @explMembers@, but always return @Filter@ for @explGet@.
+--   Can be used in cmap as @cmap $ \(Filter :: Filter a) -> b@.
+--   Since the above can be written more consicely as @cmap $ \(_ :: a) -> b@, it is rarely directly.
+--   More interestingly, we can define reusable filters like @movables = Filter :: Filter (Position, Velocity)@.
 data Filter c = Filter deriving (Eq, Show)
+
+-- Pseudostore for 'Filter'.
 newtype FilterStore s = FilterStore s
 
 instance Component c => Component (Filter c) where
@@ -145,7 +165,12 @@ instance ExplGet s => ExplGet (FilterStore s) where
   explGet _ _ = return Filter
   explExists (FilterStore s) ety = explExists s ety
 
--- | Pseudostore used to produce components of type @Entity@
+instance ExplMembers s => ExplMembers (FilterStore s) where
+  explMembers (FilterStore s) = explMembers s
+
+-- | Pseudostore used to produce components of type 'Entity'.
+-- Always returns @True@ for @explExists@, and echoes back the entity argument for @explGet@.
+-- Used in e.g. @cmap $ \(a, ety :: Entity) -> b@ to access the current entity.
 data EntityStore = EntityStore
 instance Component Entity where
   type Storage Entity = EntityStore