+++ src/Apecs/Stores.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DataKinds             #-}
+{-# LANGUAGE LambdaCase            #-}
 {-# LANGUAGE FlexibleContexts      #-}
 {-# LANGUAGE FlexibleInstances     #-}
 {-# LANGUAGE KindSignatures        #-}
@@ -54,35 +55,37 @@ instance ExplMembers (Map c) where
 -- | A Unique contains zero or one component.
 --   Writing to it overwrites both the previous component and its owner.
 --   Its main purpose is to be a @Map@ optimized for when only ever one component inhabits it.
-data Unique c = Unique (IORef Int) (IORef c)
+data Unique c = Unique (IORef (Maybe (Int, c)))
 type instance Elem (Unique c) = c
 instance ExplInit (Unique c) where
-  explInit = Unique <$> newIORef (-1) <*> newIORef (error "Uninitialized Unique value")
+  explInit = Unique <$> newIORef Nothing
 
 instance ExplGet (Unique c) where
   {-# INLINE explGet #-}
-  explGet (Unique _ cref) _ = readIORef cref
+  explGet (Unique ref) _ = readIORef ref >>= return . \case
+    Nothing -> error "Reading empty Unique"
+    Just (_, c)  -> c
   {-# INLINE explExists #-}
-  explExists (Unique eref _) ety = (==ety) <$> readIORef eref
+  explExists (Unique ref) ety = maybe False ((==ety) . fst) <$> readIORef ref
 
 instance ExplSet (Unique c) where
   {-# INLINE explSet #-}
-  explSet (Unique eref cref) ety x = writeIORef eref ety >> writeIORef cref x
+  explSet (Unique ref) ety c = writeIORef ref (Just (ety, c))
 
 instance ExplDestroy (Unique c) where
   {-# INLINE explDestroy #-}
-  explDestroy (Unique eref _) ety = do e <- readIORef eref; when (e==ety) (writeIORef eref (-1))
+  explDestroy (Unique ref) ety = writeIORef ref Nothing
 
 instance ExplMembers (Unique c) where
   {-# INLINE explMembers #-}
-  explMembers (Unique eref _) = f <$> readIORef eref
-    where f (-1) = mempty
-          f x    = U.singleton x
+  explMembers (Unique ref) = readIORef ref >>= return . \case
+    Nothing -> mempty
+    Just (ety, _) -> U.singleton ety
 
 -- | A Global contains exactly one component.
---   Initialized with 'mempty'
---   The store will return true for every existence check, but only ever gives (-1) as its inhabitant.
---   The entity argument is ignored when setting/getting a global.
+--   The initial value is 'mempty' from the component's 'Monoid' instance.
+--   When operating on a global, any entity arguments are ignored.
+--   For example, we can get a global component with @get 0@ or @get 1@ or even @get undefined@.
 newtype Global c = Global (IORef c)
 type instance Elem (Global c) = c
 instance Monoid c => ExplInit (Global c) where
@@ -96,21 +99,32 @@ instance ExplSet (Global c) where
   explSet (Global ref) _ c = writeIORef ref c
 
 -- | An empty type class indicating that the store behaves like a regular map, and can therefore safely be cached.
+--   An example of a store that cannot be cached is 'Unique'.
 class Cachable s
 instance Cachable (Map s)
 instance (KnownNat n, Cachable s) => Cachable (Cache n s)
 
 -- | A cache around another store.
+--   Caches store their members in a fixed-size vector, so operations run in O(1).
+--   Caches can provide huge performance boosts, especially for large numbers of components.
+--   The cache size is given as a type-level argument.
+--
 --   Note that iterating over a cache is linear in cache size, so sparsely populated caches might actually decrease performance.
+--   In general, the exact size of the cache does not matter as long as it reasonably approximates the number of components present.
+--
+--   The cache uses entity (-1) to internally represent missing entities, so be wary when manually manipulating entities.
 data Cache (n :: Nat) s =
   Cache Int (UM.IOVector Int) (VM.IOVector (Elem s)) s
+
+cacheMiss = error "Cache miss!"
+
 type instance Elem (Cache n s) = Elem s
 
 instance (ExplInit s, KnownNat n, Cachable s) => ExplInit (Cache n s) where
   explInit = do
     let n = fromIntegral$ natVal (Proxy @n)
     tags <- UM.replicate n (-1)
-    cache <- VM.new n
+    cache <- VM.replicate n cacheMiss
     child <- explInit
     return (Cache n tags cache child)
 
@@ -137,10 +151,13 @@ instance ExplSet s => ExplSet (Cache n s) where
     VM.unsafeWrite cache index x
 
 instance ExplDestroy s => ExplDestroy (Cache n s) where
-  explDestroy (Cache n tags _ s) ety = do
+  explDestroy (Cache n tags cache s) ety = do
+    let index = ety `rem` n
     tag <- UM.unsafeRead tags (ety `rem` n)
     if tag == ety
-       then UM.unsafeWrite tags (ety `rem` n) (-1)
+       then do
+         UM.unsafeWrite tags  index (-1)
+         VM.unsafeWrite cache index cacheMiss
        else explDestroy s ety
 
 instance ExplMembers s => ExplMembers (Cache n s) where